Refactor tests  🐿 v2.5.16
