+++ test/integration/main.spec.js
@@ -1,10 +1,10 @@
 /*global expect, sinon */
-import * as client from '../main';
+import * as client from '../../main';
 
 // require rather than import because it has to be required
 // in src to avoid circular dep (for now). If is imported here
 // but required in src then sinon fails to stub the methods
-const utils = require('../src/utils/sampleUsers');
+const utils = require('../../src/utils/sampleUsers');
 
 describe('client', () => {
 	describe('register', () => {
