+++ test/unit/caches/fonts.spec.js
@@ -1,6 +1,6 @@
 import { proxy } from 'proxyrequire';
-import { expect, sinon } from './setup';
-import { sw as precacheConfig} from '../../config/precache';
+import { expect, sinon } from '../setup';
+import { sw as precacheConfig} from '../../../config/precache';
 import makeServiceWorkerEnv from 'service-worker-mock';
 
 // Mock the service worker in the global scope to access `self` object.
@@ -22,7 +22,7 @@ describe('Fonts cache', () => {
 
 	beforeEach(() => {
 		routerStub.get = sinon.spy();
-		fontsCache = proxy(() => require('../../src/caches/fonts'), {
+		fontsCache = proxy(() => require('../../../src/caches/fonts'), {
 			'../utils/router': routerStub,
 			'../utils/precache': precacheStub
 		});
